1. Move debug and non-debug builds into a single directory, _cxxInteral is called _cxxInternal_debug in the debug build 2. Migrate from deprecated optparser to argparser in woo and woo-batch.py.in 3. Parse some options in wooOptions already, so that they are ready for "import woo" 4. Make woo importable from scripts, including all initialization; options might be set via wooOptions 5. Add --quirks arg for hardware-specific workarounds 6. Don't call Renderer::ini from when OpenGLManager, since if DISPLAY is not defined, glut would exit with error at startup

+++ SConstruct
@@ -189,7 +189,7 @@ if not env.has_key('version'): env['version']=env['realVersion']
 
 env['SUFFIX']=('-'+env['version'] if len(env['version'])>0 else '')+env['variant']
 env['SUFFIX_DBG']=env['SUFFIX']+('' if not env['debug'] else '/dbg')
-env['LIBDIR']='$PREFIX/lib/woo$SUFFIX_DBG'
+env['LIBDIR']='$PREFIX/lib/woo$SUFFIX'
 env['RESOURCEDIR']='$PREFIX/share/woo$SUFFIX'
 print "Woo version is `%s' (%s), installed files will be suffixed with `%s'."%(env['version'],env['realVersion'],env['SUFFIX'])
 buildDir=os.path.abspath(env.subst('$buildPrefix/build$SUFFIX_DBG'))
@@ -571,7 +571,7 @@ if not COMMAND_LINE_TARGETS:
 			# do not delete python-optimized files and symbolic links (lib_gts__python-module.so, for instance)
 			if ff not in toInstall and not ff.endswith('.pyo') and not ff.endswith('.pyc') and not os.path.islink(ff) and not os.path.basename(ff).startswith('.nfs'):
 				# HACK
-				if ff.endswith('/_cxxInternal.so'): continue
+				if ff.endswith('/_cxxInternal.so') or ff.endswith('_cxxInternal_debug.so'): continue
 				print "Deleting extra plugin", ff
 				os.remove(ff)
